stack + docker
Add the docker client to the stack image. I kind of want to add stack to
the docker image instead, so it can use stack path to locate the
executable, but since docker:17.12 is an alpine-based image, the easier
way to get a container with both stack and docker is to add docker to
the stack image.