use fpco/stack-build (--allow-different-user)
The workspace is owned by the jenkins user. Similarly, the git clone
step runs as jenkins user. But the stack steps run as root, and stack
complains that you don't own the project and/or parent directory, and
aborts to protect permissions.

The --allow-different-user flag lets you get around this protection,
meaning I don't have to use a custom docker image that runs as the
jenkins user, but can instead use the vanilla fpco/stack-build image
that runs as root. This is good because it's the image that most of our
projects specify (in stack.yaml) as the build environment to use when
stack's docker integration is in effect. Since I'm preventing stack from
leveraging this docker integration (since we're already in a container),
it's a good idea to use the same image that stack would've used anyway.